+++ frontend/src/interfaces/User.ts
@@ -4,5 +4,13 @@ export interface User {
   readonly id: id;
   readonly name: string;
   readonly createdAt?: string;
+  status: UserStatus;
   animal: string;
+}
+
+export enum UserStatus {
+  Enabled = 'enabled',
+  Disabled = 'disabled',
+  Deleted = 'deleted',
+  SingleSignOn = 'single-sign-on',
 }